Remote Data and Authentication - Lab
Finished CookBook - Part 3

+++ 1. JS Advanced Module/2. JS Applications/03. Remote Data and Authentication/01. Lab/01. CookBook - part 2/src/app.js
@@ -28,6 +28,11 @@ async function logout(e) {
                 }
         });
 
+        if(res.ok == false) {
+            const error = await res.json();
+            throw error;
+        }
+
         sessionStorage.clear();
         window.location = 'index.html'
 
@@ -98,7 +103,8 @@ window.addEventListener('load', async () => {
 
 });
 
-
+// template for creating recipe view;
+//type = string ; attributes = {}, any
 function createRecipe(type, attributes, ...content) {
 
     const element = document.createElement(type);
